Fixed Print Range Disabled on XPS Document Viewer
I was able to get the print range function to work.  I overrode the
default document viewer toolbar and defined a custom toolbar template in
WPF which called the print function.
The key to printing a range was to implement a wrapper of the
DocumentPaginator class which enabled the application to correctly
assign the range of page(s) to print.
When the user presses the <Print> button, the print dialog  page range
is enabled and the default value assigned is the current page being
displayed in the document viewer.
I also added document viewer current page  and total pages to display at
the bottom of the document viewer dialog.