Fixed issue with creating inputs for multiple variants
When running fvs_input.Start() on the second variant, it checked if m_ado.m_OleDbConnection was still open, but it was disposed of and set to null by the previous invocation of CreateFVSTreeInit. Now it checks for null before connection state.

Also, when working with multiple variants, the FVS_*Init and SpcdChange work tables needed to be deleted.